reworkign structure
